Move total calculations out to fix cyclical imports
