+++ test/test-utils/db-helpers/listings.ts
@@ -25,32 +25,6 @@ const allDays: string[] = [
   "Sunday",
 ];
 
-/**
- * A listing with one booking, then its stored running totals pushed off that
- * truth. The starting point for any test about drift: the booking makes the
- * recounted totals 1 and 1, and `stored` is what the listing wrongly claims.
- */
-export const createListingWithDriftedTotals = async (
-  stored: { booked_quantity: number; tickets_count: number } = {
-    booked_quantity: 9,
-    tickets_count: 5,
-  },
-): Promise<Listing> => {
-  const listing = await createTestListing({ maxAttendees: 100 });
-  // Imported here rather than at the top, because the attendee helpers import
-  // this module.
-  const { createTestAttendee } = await import("./attendees.ts");
-  await createTestAttendee(
-    listing.id,
-    listing.slug,
-    "Counted Person",
-    "counted@example.com",
-  );
-  const { listingAggregates } = await import("#db/listings/aggregates.ts");
-  await listingAggregates.update(listing.id, stored);
-  return listing;
-};
-
 /** A minute-precision ISO timestamp one minute in the past — always already
  * closed. The reference `closesAt` value for "registration closed" tests. */
 export const pastCloseTime = (): string =>
